fix: add get-open-meteo-weather-js alongside get-weather-js
