+++ examples/get-open-meteo-weather-js/.gitignore
@@ -1,0 +1,5 @@
+node_modules
+dist
+*.wasm
+pnpm-lock.yaml
+out-dir
