+++ examples/get-open-meteo-weather-js/Justfile
@@ -1,0 +1,3 @@
+build:
+    npm install 
+    npm run build:component
