+++ examples/get-open-meteo-weather-js/package.json
@@ -1,0 +1,13 @@
+{
+    "type": "module",
+    "dependencies": {
+        "@bytecodealliance/componentize-js": "0.18.1",
+        "@bytecodealliance/jco": "^1.11.1"
+    },
+    "scripts": {
+        "build:component": "jco componentize -w ./wit weather.js -o weather.wasm",
+        "transpile": "jco transpile weather.wasm -o out-dir",
+        "debug": "node use-weather.js"
+    },
+    "license": "Apache-2.0"
+}
