+++ examples/get-open-meteo-weather-js/policy.yaml
@@ -3,7 +3,5 @@ description: "Permission policy for wassette weather demo"
 permissions:
   network:
     allow:
-      - host: "api.openweathermap.org"
-  environment:
-    allow:
-      - key: "OPENWEATHER_API_KEY"
+      - host: "geocoding-api.open-meteo.com"
+      - host: "api.open-meteo.com"
