+++ examples/get-open-meteo-weather-js/wit/mcp.wit
@@ -1,9 +1,7 @@
 package mossaka:mcp@0.1.0;
 
 /// Weather information services
-world weather-mcp {
-    import wasi:config/store@0.2.0-draft;
-    
+world weather-mcp {    
     /// Get current weather information for a specified city
     export get-weather: func(city: string) -> result<string, string>;
 }
